moved interpolators to `interpolator.py`
